// src/enums/Console.ts
export enum Console {
  cleanUp = '2',
  
  removeStyles = '0',
  highlight = '1',
  grey = '2',
  italic = '3',
  underline = '4',
  twinkle = '5',
  reverse = '7',
  hidden = '8',
}

export enum Color {
  bBlack = '40',
  bDarkRed = '41',
  bGreen = '42',
  bYellow = '43',
  bBlue = '44',
  bPurple = '45',
  bDarkGreen = '46',
  bWhite = '47',

  fBlack = '30',
  fRed = '31',
  fGreen = '32',
  fYellow = '33',
  fBlue = '34',
  fPurple = '35',
  fDarkGreen = '36',
  fWhite = '37'
}